moved POST /requests to requests.ts
