+++ apps/backend/requests/create_request.rest
@@ -1,9 +1,10 @@
 # Bob's Neighborhood Id is 1
-POST http://localhost:3001/api/neighborhoods/1/requests
+POST http://localhost:3001/api/requests
 Content-Type: application/json
 Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6ImJvYjEyMzQiLCJpZCI6MSwiaWF0IjoxNjg2NzUxMDM3LCJleHAiOjE2OTAzNTEwMzd9.eohJmSkgucwCKC-jXeM_UU8EPJ4BAjSQbTXuWLh4RXA
 
 {
   "title": "foofoo",
-  "content": "barbar"
+  "content": "barbar",
+  "neighborhoodId": 1
 }
