+++ apps/backend/src/tests/neighborhood_api.test.ts
@@ -667,237 +667,175 @@ describe('Tests for getting requests associated with a n-hood GET /neighborhoods
   });
 });
 
-// describe('Test for getting a single request at GET /neighborhoods/:id/requests/:id', () => {
-//   beforeAll(async () => {
+// describe('Tests for creating a new request at POST /requests', () => {
+//   beforeEach(async () => {
 //     await seed();
 //   });
 
-//   test('GET /neighborhoods/:nId/requests/:rId fails when no authorization header present', async () => {
-//     const getResponse: Response = await api.get(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests/${MIKES_REQUEST_ID}`);
+//   test('POST /:nId/requests fails when no token exists', async () => {
+//     const numberOfInitialRequests = await testHelpers.getNumberOfRequests();
 
-//     expect(getResponse.status).toEqual(401);
-//     expect(getResponse.body.error).toEqual('user not signed in');
+//     const postResponse: Response = await api.post(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests`);
+
+//     const numberOfFinalRequests = await testHelpers.getNumberOfRequests();
+
+//     expect(postResponse.status).toEqual(401);
+//     expect(postResponse.body.error).toEqual('user not signed in');
+
+//     expect(numberOfInitialRequests).toEqual(numberOfFinalRequests);
 //   });
 
-//   test('GET /neighborhoods/:nId/requests/:rId/ fails when user not a member of neighborhood', async () => {
+//   test('POST /:nId/requests fails when token invalid', async () => {
+//     const numberOfInitialRequests = await testHelpers.getNumberOfRequests();
+
+//     const postResponse: Response = await api
+//       .post(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests`)
+//       .set('Authorization', 'WrongToken');
+
+//     const numberOfFinalRequests = await testHelpers.getNumberOfRequests();
+
+//     expect(postResponse.status).toEqual(401);
+//     expect(postResponse.body.error).toEqual('user not signed in');
+//     expect(numberOfInitialRequests).toEqual(numberOfFinalRequests);
+//   });
+
+//   test('POST /:nId/requests fails when data missing', async () => {
 //     const loginResponse = await loginUser(BOBS_LOGIN_DATA);
 //     const { token } = loginResponse.body;
 
-//     const getResponse: Response = await api
-//       .get(`/api/neighborhoods/${ANTONINAS_NHOOD_ID}/requests/${MIKES_REQUEST_ID}`)
-//       .set('Authorization', `Bearer ${token}`);
+//     const numberOfInitialRequests = await testHelpers.getNumberOfRequests();
 
-//     expect(getResponse.status).toEqual(401);
-//     expect(getResponse.body.error).toEqual('user not a member of neighborhood');
+//     // no content
+//     const response1 = await api
+//       .post(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests`)
+//       .set('Authorization', `Bearer ${token}`)
+//       .send({ title: 'foo' })
+//       .expect(400)
+//       .expect('Content-Type', /application\/json/);
+
+//     let numberOfCurrentRequests = await testHelpers.getNumberOfRequests();
+//     expect(numberOfCurrentRequests).toEqual(numberOfInitialRequests);
+//     expect(response1.body.error).toEqual('title or content missing or invalid');
+
+//     // no title
+//     const response2 = await api
+//       .post(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests`)
+//       .set('Authorization', `Bearer ${token}`)
+//       .send({ neighborhood_id: BOBS_NHOOD_ID, content: 'foo' })
+//       .expect(400)
+//       .expect('Content-Type', /application\/json/);
+
+//     numberOfCurrentRequests = await testHelpers.getNumberOfRequests();
+//     expect(numberOfCurrentRequests).toEqual(numberOfInitialRequests);
+//     expect(response2.body.error).toEqual('title or content missing or invalid');
 //   });
 
-//   test('GET /neighborhood/:nId/requests/:rId fails when neighborhoodId invalid', async () => {
+//   test('POST /:nId/requests fails when neighborhood does not exist', async () => {
 //     const loginResponse = await loginUser(BOBS_LOGIN_DATA);
 //     const { token } = loginResponse.body;
 
-//     const getResponse: Response = await api
-//       .get(`/api/neighborhoods/foo/requests/${MIKES_REQUEST_ID}`)
-//       .set('Authorization', `Bearer ${token}`);
+//     const numberOfInitialRequests = await testHelpers.getNumberOfRequests();
 
-//     expect(getResponse.status).toEqual(400);
+//     const response = await api
+//       .post(`/api/neighborhoods/${INVALID_NHOOD_ID}/requests`)
+//       .set('Authorization', `Bearer ${token}`)
+//       .send({ title: 'foofoo', content: 'barbar' })
+//       .expect(400)
+//       .expect('Content-Type', /application\/json/);
+
+//     const numberOfFinalRequests = await testHelpers.getNumberOfRequests();
+//     expect(numberOfFinalRequests).toEqual(numberOfInitialRequests);
+//     expect(response.body.error).toBe('Neighborhood does not exist');
 //   });
 
-//   test('GET /neighborhood/:nId/requests/:rId fails when request is not associated with nhood', async () => {
+//   test('POST /request fails when content invalid', async () => {
 //     const loginResponse = await loginUser(BOBS_LOGIN_DATA);
 //     const { token } = loginResponse.body;
 
-//     const NOT_MIKES_REQUEST_ID = 2;
-//     const getResponse: Response = await api
-//       .get(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests/${NOT_MIKES_REQUEST_ID}`)
-//       .set('Authorization', `Bearer ${token}`);
+//     const numberOfInitialRequests = await testHelpers.getNumberOfRequests();
 
-//     expect(getResponse.status).toEqual(400);
-//     expect(getResponse.body.error).toEqual('request not associated with the neighborhood');
+//     const INVALID_TITLE = 'foo';
+//     const response = await api
+//       .post(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests`)
+//       .set('Authorization', `Bearer ${token}`)
+//       .send({ title: INVALID_TITLE, content: 'barbar' })
+//       .expect(400)
+//       .expect('Content-Type', /application\/json/);
+
+//     const numberOfFinalRequests = await testHelpers.getNumberOfRequests();
+//     expect(numberOfFinalRequests).toEqual(numberOfInitialRequests);
+//     expect(response.body.error).toBe('Invalid title');
 //   });
 
-//   test('GET /neighborhoods/:nId/requests/:rId/ fails works with valid data', async () => {
+//   test('POST /request fails when user not a member of the neighborhood', async () => {
 //     const loginResponse = await loginUser(BOBS_LOGIN_DATA);
 //     const { token } = loginResponse.body;
 
-//     const getResponse: Response = await api
-//       .get(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests/${MIKES_REQUEST_ID}`)
-//       .set('Authorization', `Bearer ${token}`);
+//     const numberOfInitialRequests = await testHelpers.getNumberOfRequests();
 
-//     expect(getResponse.status).toEqual(200);
-//     expect(getResponse.body.user_id).toBe(MIKES_USER_ID);
-//     expect(getResponse.body.title).toBe(MIKES_REQUEST_TITLE);
+//     const response = await api
+//       .post(`/api/neighborhoods/${ANTONINAS_NHOOD_ID}/requests`)
+//       .set('Authorization', `Bearer ${token}`)
+//       .send({ title: 'foofoo', content: 'barbar' })
+//       .expect(400)
+//       .expect('Content-Type', /application\/json/);
+
+//     const numberOfCurrentRequests = await testHelpers.getNumberOfRequests();
+//     expect(numberOfCurrentRequests).toEqual(numberOfInitialRequests);
+//     expect(response.body.error).toBe('User is not a member of neighborhood');
+//   });
+
+//   test('POST /request succeeds with valid data', async () => {
+//     const loginResponse = await loginUser(BOBS_LOGIN_DATA);
+//     const { token } = loginResponse.body;
+
+//     const numberOfInitialRequests = await testHelpers.getNumberOfRequests();
+
+//     const initialRequestsAssociatedWithBob = await testHelpers
+//       .getRequestsAssociatedWithUser(BOBS_USER_ID);
+//     const numberOfInitialRequestsAssociatedWithBob = initialRequestsAssociatedWithBob.length;
+
+//     const initialRequestsAssociatedWithNeighborhood = await testHelpers
+//       .getRequestsAssociatedWithNeighborhood(BOBS_NHOOD_ID);
+//     const numInitialRequestsAssociatedWithNeighborhood = initialRequestsAssociatedWithNeighborhood
+//       .length;
+
+//     const response = await api
+//       .post(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests`)
+//       .set('Authorization', `Bearer ${token}`)
+//       .send({ title: 'foofoo', content: 'barbar' })
+//       .expect(201)
+//       .expect('Content-Type', /application\/json/);
+
+//     const numberOfFinalRequests = await testHelpers.getNumberOfRequests();
+
+//     const finalRequestsAssociatedWithBob = await testHelpers
+//       .getRequestsAssociatedWithUser(BOBS_USER_ID);
+//     const numberOfFinalRequestsAssociatedWithBob = finalRequestsAssociatedWithBob.length;
+//     const bobsRequestContentsAfterCreation = finalRequestsAssociatedWithBob.map(r => r.content);
+
+//     const finalRequestsAssociatedWithNeighborhood = await testHelpers
+//       .getRequestsAssociatedWithNeighborhood(BOBS_NHOOD_ID);
+//     const numberOfFinalRequestsAssociatedWithNeighborhood = finalRequestsAssociatedWithNeighborhood
+//       .length;
+//     const neighborhoodsRequestTitlesAfterCreation = finalRequestsAssociatedWithNeighborhood
+//       .map(r => r.title);
+
+//     expect(response.body.neighborhood_id).toEqual(BOBS_NHOOD_ID);
+//     expect(response.body.user_id).toEqual(BOBS_USER_ID);
+//     expect(response.body.title).toEqual('foofoo');
+//     expect(response.body.content).toEqual('barbar');
+//     expect(response.body.status).toEqual('OPEN');
+//     expect(response.body.time_created).toBeDefined();
+
+//     expect(numberOfFinalRequests).toEqual(numberOfInitialRequests + 1);
+
+//     expect(numberOfFinalRequestsAssociatedWithBob)
+//       .toEqual(numberOfInitialRequestsAssociatedWithBob + 1);
+//     expect(bobsRequestContentsAfterCreation).toContain('barbar');
+
+//     expect(numberOfFinalRequestsAssociatedWithNeighborhood)
+//       .toEqual(numInitialRequestsAssociatedWithNeighborhood + 1);
+//     expect(neighborhoodsRequestTitlesAfterCreation).toContain('foofoo');
 //   });
 // });
-
-describe('Tests for creating a new request at POST /requests', () => {
-  beforeEach(async () => {
-    await seed();
-  });
-
-  test('POST /:nId/requests fails when no token exists', async () => {
-    const numberOfInitialRequests = await testHelpers.getNumberOfRequests();
-
-    const postResponse: Response = await api.post(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests`);
-
-    const numberOfFinalRequests = await testHelpers.getNumberOfRequests();
-
-    expect(postResponse.status).toEqual(401);
-    expect(postResponse.body.error).toEqual('user not signed in');
-
-    expect(numberOfInitialRequests).toEqual(numberOfFinalRequests);
-  });
-
-  test('POST /:nId/requests fails when token invalid', async () => {
-    const numberOfInitialRequests = await testHelpers.getNumberOfRequests();
-
-    const postResponse: Response = await api
-      .post(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests`)
-      .set('Authorization', 'WrongToken');
-
-    const numberOfFinalRequests = await testHelpers.getNumberOfRequests();
-
-    expect(postResponse.status).toEqual(401);
-    expect(postResponse.body.error).toEqual('user not signed in');
-    expect(numberOfInitialRequests).toEqual(numberOfFinalRequests);
-  });
-
-  test('POST /:nId/requests fails when data missing', async () => {
-    const loginResponse = await loginUser(BOBS_LOGIN_DATA);
-    const { token } = loginResponse.body;
-
-    const numberOfInitialRequests = await testHelpers.getNumberOfRequests();
-
-    // no content
-    const response1 = await api
-      .post(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests`)
-      .set('Authorization', `Bearer ${token}`)
-      .send({ title: 'foo' })
-      .expect(400)
-      .expect('Content-Type', /application\/json/);
-
-    let numberOfCurrentRequests = await testHelpers.getNumberOfRequests();
-    expect(numberOfCurrentRequests).toEqual(numberOfInitialRequests);
-    expect(response1.body.error).toEqual('title or content missing or invalid');
-
-    // no title
-    const response2 = await api
-      .post(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests`)
-      .set('Authorization', `Bearer ${token}`)
-      .send({ neighborhood_id: BOBS_NHOOD_ID, content: 'foo' })
-      .expect(400)
-      .expect('Content-Type', /application\/json/);
-
-    numberOfCurrentRequests = await testHelpers.getNumberOfRequests();
-    expect(numberOfCurrentRequests).toEqual(numberOfInitialRequests);
-    expect(response2.body.error).toEqual('title or content missing or invalid');
-  });
-
-  test('POST /:nId/requests fails when neighborhood does not exist', async () => {
-    const loginResponse = await loginUser(BOBS_LOGIN_DATA);
-    const { token } = loginResponse.body;
-
-    const numberOfInitialRequests = await testHelpers.getNumberOfRequests();
-
-    const response = await api
-      .post(`/api/neighborhoods/${INVALID_NHOOD_ID}/requests`)
-      .set('Authorization', `Bearer ${token}`)
-      .send({ title: 'foofoo', content: 'barbar' })
-      .expect(400)
-      .expect('Content-Type', /application\/json/);
-
-    const numberOfFinalRequests = await testHelpers.getNumberOfRequests();
-    expect(numberOfFinalRequests).toEqual(numberOfInitialRequests);
-    expect(response.body.error).toBe('Neighborhood does not exist');
-  });
-
-  test('POST /request fails when content invalid', async () => {
-    const loginResponse = await loginUser(BOBS_LOGIN_DATA);
-    const { token } = loginResponse.body;
-
-    const numberOfInitialRequests = await testHelpers.getNumberOfRequests();
-
-    const INVALID_TITLE = 'foo';
-    const response = await api
-      .post(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests`)
-      .set('Authorization', `Bearer ${token}`)
-      .send({ title: INVALID_TITLE, content: 'barbar' })
-      .expect(400)
-      .expect('Content-Type', /application\/json/);
-
-    const numberOfFinalRequests = await testHelpers.getNumberOfRequests();
-    expect(numberOfFinalRequests).toEqual(numberOfInitialRequests);
-    expect(response.body.error).toBe('Invalid title');
-  });
-
-  test('POST /request fails when user not a member of the neighborhood', async () => {
-    const loginResponse = await loginUser(BOBS_LOGIN_DATA);
-    const { token } = loginResponse.body;
-
-    const numberOfInitialRequests = await testHelpers.getNumberOfRequests();
-
-    const response = await api
-      .post(`/api/neighborhoods/${ANTONINAS_NHOOD_ID}/requests`)
-      .set('Authorization', `Bearer ${token}`)
-      .send({ title: 'foofoo', content: 'barbar' })
-      .expect(400)
-      .expect('Content-Type', /application\/json/);
-
-    const numberOfCurrentRequests = await testHelpers.getNumberOfRequests();
-    expect(numberOfCurrentRequests).toEqual(numberOfInitialRequests);
-    expect(response.body.error).toBe('User is not a member of neighborhood');
-  });
-
-  test('POST /request succeeds with valid data', async () => {
-    const loginResponse = await loginUser(BOBS_LOGIN_DATA);
-    const { token } = loginResponse.body;
-
-    const numberOfInitialRequests = await testHelpers.getNumberOfRequests();
-
-    const initialRequestsAssociatedWithBob = await testHelpers
-      .getRequestsAssociatedWithUser(BOBS_USER_ID);
-    const numberOfInitialRequestsAssociatedWithBob = initialRequestsAssociatedWithBob.length;
-
-    const initialRequestsAssociatedWithNeighborhood = await testHelpers
-      .getRequestsAssociatedWithNeighborhood(BOBS_NHOOD_ID);
-    const numInitialRequestsAssociatedWithNeighborhood = initialRequestsAssociatedWithNeighborhood
-      .length;
-
-    const response = await api
-      .post(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests`)
-      .set('Authorization', `Bearer ${token}`)
-      .send({ title: 'foofoo', content: 'barbar' })
-      .expect(201)
-      .expect('Content-Type', /application\/json/);
-
-    const numberOfFinalRequests = await testHelpers.getNumberOfRequests();
-
-    const finalRequestsAssociatedWithBob = await testHelpers
-      .getRequestsAssociatedWithUser(BOBS_USER_ID);
-    const numberOfFinalRequestsAssociatedWithBob = finalRequestsAssociatedWithBob.length;
-    const bobsRequestContentsAfterCreation = finalRequestsAssociatedWithBob.map(r => r.content);
-
-    const finalRequestsAssociatedWithNeighborhood = await testHelpers
-      .getRequestsAssociatedWithNeighborhood(BOBS_NHOOD_ID);
-    const numberOfFinalRequestsAssociatedWithNeighborhood = finalRequestsAssociatedWithNeighborhood
-      .length;
-    const neighborhoodsRequestTitlesAfterCreation = finalRequestsAssociatedWithNeighborhood
-      .map(r => r.title);
-
-    expect(response.body.neighborhood_id).toEqual(BOBS_NHOOD_ID);
-    expect(response.body.user_id).toEqual(BOBS_USER_ID);
-    expect(response.body.title).toEqual('foofoo');
-    expect(response.body.content).toEqual('barbar');
-    expect(response.body.status).toEqual('OPEN');
-    expect(response.body.time_created).toBeDefined();
-
-    expect(numberOfFinalRequests).toEqual(numberOfInitialRequests + 1);
-
-    expect(numberOfFinalRequestsAssociatedWithBob)
-      .toEqual(numberOfInitialRequestsAssociatedWithBob + 1);
-    expect(bobsRequestContentsAfterCreation).toContain('barbar');
-
-    expect(numberOfFinalRequestsAssociatedWithNeighborhood)
-      .toEqual(numInitialRequestsAssociatedWithNeighborhood + 1);
-    expect(neighborhoodsRequestTitlesAfterCreation).toContain('foofoo');
-  });
-});
