+++ apps/backend/src/tests/requests_api.test.ts
@@ -4,7 +4,7 @@ import { Response } from 'supertest';
 import app from '../app';
 import prismaClient from '../../prismaClient';
 import seed from './seed';
-// import testHelpers from './testHelpers';
+import testHelpers from './testHelpers';
 import { LoginData, UpdateRequestData } from '../types';
 
 const supertest = require('supertest'); // eslint-disable-line
@@ -35,6 +35,8 @@ const MIKES_REQUEST_ID = 1;
 const RADUS_REQUEST_ID = 2;
 const MIKES_USER_ID = 5;
 const MIKES_REQUEST_TITLE = 'Help moving furniture in apartment';
+const INVALID_NHOOD_ID = 1000;
+const ANTONINAS_NHOOD_ID = 2;
 
 const loginUser = async (loginData: LoginData) => {
   const loginResponse = await api
@@ -364,5 +366,190 @@ describe('Test for getting a single request at GET /requests/:id', () => {
     expect(getResponse.status).toEqual(200);
     expect(getResponse.body.user_id).toBe(MIKES_USER_ID);
     expect(getResponse.body.title).toBe(MIKES_REQUEST_TITLE);
+  });
+});
+
+describe('Tests for creating a new request at POST /requests', () => {
+  beforeEach(async () => {
+    await seed();
+  });
+
+  test('POST /requests fails when no token exists', async () => {
+    const numberOfInitialRequests = await testHelpers.getNumberOfRequests();
+
+    const postResponse: Response = await api.post('/api/requests');
+
+    const numberOfFinalRequests = await testHelpers.getNumberOfRequests();
+
+    expect(postResponse.status).toEqual(401);
+    expect(postResponse.body.error).toEqual('user not signed in');
+
+    expect(numberOfInitialRequests).toEqual(numberOfFinalRequests);
+  });
+
+  test('POST /requests fails when token invalid', async () => {
+    const numberOfInitialRequests = await testHelpers.getNumberOfRequests();
+
+    const postResponse: Response = await api
+      .post('/api/requests')
+      .set('Authorization', 'WrongToken');
+
+    const numberOfFinalRequests = await testHelpers.getNumberOfRequests();
+
+    expect(postResponse.status).toEqual(401);
+    expect(postResponse.body.error).toEqual('user not signed in');
+    expect(numberOfInitialRequests).toEqual(numberOfFinalRequests);
+  });
+
+  test('POST /requests fails when data missing', async () => {
+    const loginResponse = await loginUser(BOBS_LOGIN_DATA);
+    const { token } = loginResponse.body;
+
+    const numberOfInitialRequests = await testHelpers.getNumberOfRequests();
+
+    // no content
+    const response1 = await api
+      .post('/api/requests')
+      .set('Authorization', `Bearer ${token}`)
+      .send({ title: 'foo', neighborhoodId: BOBS_NHOOD_ID })
+      .expect(400)
+      .expect('Content-Type', /application\/json/);
+
+    let numberOfCurrentRequests = await testHelpers.getNumberOfRequests();
+    expect(numberOfCurrentRequests).toEqual(numberOfInitialRequests);
+    expect(response1.body.error).toEqual('title, content or neighborhoodId missing or invalid');
+
+    // no title
+    const response2 = await api
+      .post('/api/requests')
+      .set('Authorization', `Bearer ${token}`)
+      .send({ neighborhoodId: BOBS_NHOOD_ID, content: 'foo' })
+      .expect(400)
+      .expect('Content-Type', /application\/json/);
+
+    numberOfCurrentRequests = await testHelpers.getNumberOfRequests();
+    expect(numberOfCurrentRequests).toEqual(numberOfInitialRequests);
+    expect(response2.body.error).toEqual('title, content or neighborhoodId missing or invalid');
+
+    // no neighborhoodId
+    const response = await api
+      .post('/api/requests')
+      .set('Authorization', `Bearer ${token}`)
+      .send({ title: 'foofoo', content: 'barbar' })
+      .expect(400)
+      .expect('Content-Type', /application\/json/);
+
+    numberOfCurrentRequests = await testHelpers.getNumberOfRequests();
+    expect(numberOfCurrentRequests).toEqual(numberOfInitialRequests);
+    expect(response.body.error).toBe('title, content or neighborhoodId missing or invalid');
+  });
+
+  test('POST /requests fails when neighborhood does not exist', async () => {
+    const loginResponse = await loginUser(BOBS_LOGIN_DATA);
+    const { token } = loginResponse.body;
+
+    const numberOfInitialRequests = await testHelpers.getNumberOfRequests();
+
+    const response = await api
+      .post('/api/requests')
+      .set('Authorization', `Bearer ${token}`)
+      .send({ title: 'foofoo', content: 'barbar', neighborhoodId: INVALID_NHOOD_ID })
+      .expect(400)
+      .expect('Content-Type', /application\/json/);
+
+    const numberOfFinalRequests = await testHelpers.getNumberOfRequests();
+    expect(numberOfFinalRequests).toEqual(numberOfInitialRequests);
+    expect(response.body.error).toBe('Neighborhood does not exist');
+  });
+
+  test('POST /request fails when content invalid', async () => {
+    const loginResponse = await loginUser(BOBS_LOGIN_DATA);
+    const { token } = loginResponse.body;
+
+    const numberOfInitialRequests = await testHelpers.getNumberOfRequests();
+
+    const INVALID_TITLE = 'foo';
+    const response = await api
+      .post('/api/requests')
+      .set('Authorization', `Bearer ${token}`)
+      .send({ title: INVALID_TITLE, content: 'barbar', neighborhoodId: BOBS_NHOOD_ID })
+      .expect(400)
+      .expect('Content-Type', /application\/json/);
+
+    const numberOfFinalRequests = await testHelpers.getNumberOfRequests();
+    expect(numberOfFinalRequests).toEqual(numberOfInitialRequests);
+    expect(response.body.error).toBe('Invalid title');
+  });
+
+  test('POST /request fails when user not a member of the neighborhood', async () => {
+    const loginResponse = await loginUser(BOBS_LOGIN_DATA);
+    const { token } = loginResponse.body;
+
+    const numberOfInitialRequests = await testHelpers.getNumberOfRequests();
+
+    const response = await api
+      .post('/api/requests')
+      .set('Authorization', `Bearer ${token}`)
+      .send({ title: 'foofoo', content: 'barbar', neighborhoodId: ANTONINAS_NHOOD_ID })
+      .expect(400)
+      .expect('Content-Type', /application\/json/);
+
+    const numberOfCurrentRequests = await testHelpers.getNumberOfRequests();
+    expect(numberOfCurrentRequests).toEqual(numberOfInitialRequests);
+    expect(response.body.error).toBe('User is not a member of neighborhood');
+  });
+
+  test('POST /request succeeds with valid data', async () => {
+    const loginResponse = await loginUser(BOBS_LOGIN_DATA);
+    const { token } = loginResponse.body;
+
+    const numberOfInitialRequests = await testHelpers.getNumberOfRequests();
+
+    const initialRequestsAssociatedWithBob = await testHelpers
+      .getRequestsAssociatedWithUser(BOBS_USER_ID);
+    const numberOfInitialRequestsAssociatedWithBob = initialRequestsAssociatedWithBob.length;
+
+    const initialRequestsAssociatedWithNeighborhood = await testHelpers
+      .getRequestsAssociatedWithNeighborhood(BOBS_NHOOD_ID);
+    const numInitialRequestsAssociatedWithNeighborhood = initialRequestsAssociatedWithNeighborhood
+      .length;
+
+    const response = await api
+      .post('/api/requests')
+      .set('Authorization', `Bearer ${token}`)
+      .send({ title: 'foofoo', content: 'barbar', neighborhoodId: BOBS_NHOOD_ID })
+      .expect(201)
+      .expect('Content-Type', /application\/json/);
+
+    const numberOfFinalRequests = await testHelpers.getNumberOfRequests();
+
+    const finalRequestsAssociatedWithBob = await testHelpers
+      .getRequestsAssociatedWithUser(BOBS_USER_ID);
+    const numberOfFinalRequestsAssociatedWithBob = finalRequestsAssociatedWithBob.length;
+    const bobsRequestContentsAfterCreation = finalRequestsAssociatedWithBob.map(r => r.content);
+
+    const finalRequestsAssociatedWithNeighborhood = await testHelpers
+      .getRequestsAssociatedWithNeighborhood(BOBS_NHOOD_ID);
+    const numberOfFinalRequestsAssociatedWithNeighborhood = finalRequestsAssociatedWithNeighborhood
+      .length;
+    const neighborhoodsRequestTitlesAfterCreation = finalRequestsAssociatedWithNeighborhood
+      .map(r => r.title);
+
+    expect(response.body.neighborhood_id).toEqual(BOBS_NHOOD_ID);
+    expect(response.body.user_id).toEqual(BOBS_USER_ID);
+    expect(response.body.title).toEqual('foofoo');
+    expect(response.body.content).toEqual('barbar');
+    expect(response.body.status).toEqual('OPEN');
+    expect(response.body.time_created).toBeDefined();
+
+    expect(numberOfFinalRequests).toEqual(numberOfInitialRequests + 1);
+
+    expect(numberOfFinalRequestsAssociatedWithBob)
+      .toEqual(numberOfInitialRequestsAssociatedWithBob + 1);
+    expect(bobsRequestContentsAfterCreation).toContain('barbar');
+
+    expect(numberOfFinalRequestsAssociatedWithNeighborhood)
+      .toEqual(numInitialRequestsAssociatedWithNeighborhood + 1);
+    expect(neighborhoodsRequestTitlesAfterCreation).toContain('foofoo');
   });
 });
